move private investment functions into Corporation/Actions to support testability
